modified according to review comments
