+++ examples/classification/ofa/ofa_imagenet_search_expand_phase1.json
@@ -1,20 +1,26 @@
 {
     "dataloader": {
-        "imagenet_ofa": {
-            "train_path": "IMAGENET_TRAINING_DIR",
-            "train_file": "IMAGENET_TRAINING_LABEL",
-            "valid_path": "IMAGENET_TEST_DIR",
-            "valid_file": "IMAGENET_TEST_LABEL"
+        "imagenet": {
+            "train_portion": 0.99,
+            "augment_valid": "False",
+            "colortwist": "True",
+            "train_path": "/data/imagenet/train",
+            "train_file": "/home/yuiko.sakuma/data/ilsvrc2012/imagenet_dir/train_label",
+            "valid_path": "/data/imagenet/val",
+            "valid_file": "/home/yuiko.sakuma/data/ilsvrc2012/imagenet_dir/val_label"
         }
     },
     "network": {
         "ofa": {
         "num_classes": 1000,
         "bn_param": [0.9, 1e-5],
-        "dropout": 0.1,
-        "ks_list": [3, 5, 7],
-        "expand_ratio_list": [4, 6],
-        "depth_list": [2, 3, 4]
+        "drop_rate": 0.1,
+        "op_candidates": [
+            "MB4 3x3", "MB4 5x5", "MB4 7x7", 
+            "MB6 3x3", "MB6 5x5", "MB6 7x7"
+        ],
+        "depth_candidates": [2, 3, 4],
+        "weights": "log/classification/ofa/imagenet/search/K357_E6_D234/weights.h5"
         }
     },
     "optimizer": {
@@ -35,23 +41,42 @@
     },
     "hparams": {
         "task": "expand", 
-        "teacher_path": "nnabla-nas/examples/classification/ofa/sample_params/R_K7_E6_D4_acc0.75968.h5",
-        "pretrained_path": "nnabla-nas/examples/classification/ofa/sample_params/R_K357_E6_D234_acc0.79395.h5",
-        "num_classes": 1000, 
+        "train_image_size_list": [128, 160, 192, 224],
+        "valid_image_size_list": [160, 224],
+        "valid_genotypes": [
+            [0, 0, 6, 6, 0, 0, 6, 6, 0, 0, 6, 6, 0, 0, 6, 6, 0, 0, 6, 6],
+            [2, 2, 6, 6, 2, 2, 6, 6, 2, 2, 6, 6, 2, 2, 6, 6, 2, 2, 6, 6],
+            [3, 3, 6, 6, 3, 3, 6, 6, 3, 3, 6, 6, 3, 3, 6, 6, 3, 3, 6, 6],
+            [5, 5, 6, 6, 5, 5, 6, 6, 5, 5, 6, 6, 5, 5, 6, 6, 5, 5, 6, 6],
+            [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+            [2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2],
+            [3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3],
+            [5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5]
+        ],
         "input_shapes": [
             [3, 224, 224]
         ],
         "target_shapes": [
             [1]
         ],
-        "no_decay_keys": "bn#bias",
-        "kd_ratio": 1,
+        "no_decay_keys": "bn#_b",
+        "lambda_kd": 1,
+        "teacher_network": {
+            "ofa": {
+            "num_classes": 1000,
+            "bn_param": [0.9, 1e-5],
+            "drop_rate": 0,
+            "op_candidates": ["MB6 7x7"],
+            "depth_candidates": [4],
+            "weights": "log/classification/ofa/imagenet/search/K7_E6_D4/weights.h5"
+            }
+        },
         "epoch": 25,
         "cosine_warmup_epoch": 0,
         "warmup_lr": 0.0025,
         "batch_size_train": 256,
         "batch_size_valid": 200,
-        "mini_batch_train": 32,
+        "mini_batch_train": 16,
         "mini_batch_valid": 200,
         "print_frequency": 100,
         "validation_frequency": 1
