+++ nnabla_nas/contrib/classification/ofa/modules.py
@@ -14,15 +14,63 @@
 
 from collections import OrderedDict
 
-import numpy as np
-
-import nnabla as nn
-import nnabla.functions as F
-
 from .... import module as Mo
-from .ofa_modules.common_tools import make_divisible, get_same_padding, min_divisible_value
-from .ofa_modules.pytorch_modules import SEModule, Hswish
-from .ofa_modules.my_modules import MyNetwork
+from ..base import ClassificationModel as Model
+from .ofa_modules.static_op import SEModule
+from .ofa_utils.common_tools import get_same_padding, min_divisible_value
+from .ofa_utils.common_tools import cross_entropy_loss_with_label_smoothing
+
+
+CANDIDATES = {
+    'MB3 3x3': {'ks': 3, 'expand_ratio': 3},
+    'MB3 5x5': {'ks': 5, 'expand_ratio': 3},
+    'MB3 7x7': {'ks': 7, 'expand_ratio': 3},
+    'MB4 3x3': {'ks': 3, 'expand_ratio': 4},
+    'MB4 5x5': {'ks': 5, 'expand_ratio': 4},
+    'MB4 7x7': {'ks': 7, 'expand_ratio': 4},
+    'MB6 3x3': {'ks': 3, 'expand_ratio': 6},
+    'MB6 5x5': {'ks': 5, 'expand_ratio': 6},
+    'MB6 7x7': {'ks': 7, 'expand_ratio': 6},
+    'skip_connect': {'ks': None, 'expand_ratio': None},
+}
+
+
+def candidates2subnetlist(candidates):
+    ks_list = []
+    expand_list = []
+    for candidate in candidates:
+        ks = CANDIDATES[candidate]['ks']
+        e = CANDIDATES[candidate]['expand_ratio']
+        if ks not in ks_list:
+            ks_list.append(ks)
+        if e not in expand_list:
+            expand_list.append(e)
+    return ks_list, expand_list
+
+
+def genotype2subnetlist(op_candidates, genotype):
+    op_candidates.append('skip_connect')
+    subnet_list = [op_candidates[i] for i in genotype]
+    ks_list = [CANDIDATES[subnet]['ks'] if subnet != 'skip_connect'
+               else 3 for subnet in subnet_list]
+    expand_ratio_list = [CANDIDATES[subnet]['expand_ratio'] if subnet != 'skip_connect'
+                         else 4 for subnet in subnet_list]
+    depth_list = []
+    d = 0
+    for i, subnet in enumerate(subnet_list):
+        if subnet == 'skip_connect':
+            if d > 1:
+                depth_list.append(d)
+                d = 0
+        elif d == 4:
+            depth_list.append(d)
+            d = 1
+        elif i == len(subnet_list) - 1:
+            depth_list.append(d + 1)
+        else:
+            d += 1
+    assert([d > 1 for d in depth_list])
+    return ks_list, expand_ratio_list, depth_list
 
 
 def set_layer_from_config(layer_config):
@@ -31,7 +79,6 @@ def set_layer_from_config(layer_config):
 
     name2layer = {
         ConvLayer.__name__: ConvLayer,
-        IdentityLayer.__name__: IdentityLayer,
         LinearLayer.__name__: LinearLayer,
         MBConvLayer.__name__: MBConvLayer,
         'MBInvertedConvLayer': MBConvLayer,
@@ -44,11 +91,84 @@ def set_layer_from_config(layer_config):
     return layer.build_from_config(layer_config)
 
 
-class ResidualBlock(Mo.Module):
-    def __init__(self, conv, shortcut):
-        super(ResidualBlock, self).__init__()
-        self._scope_name = f'<residualblock at {hex(id(self))}>'
+def set_bn_param(net, decay_rate, eps, **kwargs):
+    for _, m in net.get_modules():
+        if isinstance(m, Mo.BatchNormalization):
+            m._decay_rate = decay_rate
+            m._eps = eps
 
+
+def get_bn_param(net):
+    for _, m in net.get_modules():
+        if isinstance(m, Mo.BatchNormalization):
+            return {
+                'decay_rate': m._decay_rate,
+                'eps': m._eps
+            }
+
+
+class MyNetwork(Model):
+    CHANNEL_DIVISIBLE = 8
+
+    def set_bn_param(self, decay_rate, eps, **kwargs):
+        r"""Sets decay_rate and eps to batchnormalization layers.
+
+        Args:
+            decay_rate (float): Deccay rate of running mean and variance.
+            eps (float):Tiny value to avoid zero division by std.
+        """
+        set_bn_param(self, decay_rate, eps, **kwargs)
+
+    def get_bn_param(self):
+        r"""Return dict of batchnormalization params.
+
+        Returns:
+            dict: A dictionary containing decay_rate and eps of batchnormalization
+        """
+        return get_bn_param(self)
+
+    def loss(self, outputs, targets, loss_weights=None):
+        r"""Return loss computed from a list of outputs and list of targets.
+
+        Args:
+            outputs (list of nn.Variable):
+                A list of output variables computed from the model.
+            targets (list of nn.Variable):
+                A list of target variables loaded from the data.
+            loss_weights (list of float, optional):
+                A list specifying scalar coefficients to weight the loss
+                contributions of different model outputs.
+                It is expected to have a 1:1 mapping to model outputs.
+                Defaults to None.
+        Returns:
+            nn.Variable: A scalar NNabla Variable represents the loss.
+        """
+        return cross_entropy_loss_with_label_smoothing(outputs[0], targets[0])
+
+    def get_net_parameters(self, grad_only=False):
+        r"""Returns an `OrderedDict` containing architecture parameters.
+
+        Args:
+            grad_only (bool, optional): If sets to `True`, then only parameters
+                with `need_grad=True` are returned. Defaults to False.
+        Returns:
+            OrderedDict: A dictionary containing parameters.
+        """
+        p = self.get_parameters(grad_only)
+        return OrderedDict([(k, v) for k, v in p.items()])
+
+
+class ResidualBlock(Mo.Module):
+    r"""ResidualBlock layer.
+
+    Adds outputs of a convolution layer and a shortcut.
+
+    Args:
+        conv (:obj:`Module`): A convolution module.
+        shortcut (:obj:`Module`): An identity module.
+    """
+
+    def __init__(self, conv, shortcut):
         self.conv = conv
         self.shortcut = shortcut
 
@@ -61,127 +181,53 @@ class ResidualBlock(Mo.Module):
             res = self.conv(x) + self.shortcut(x)
         return res
 
-    @property
-    def module_str(self):
-        return '(%s, %s)' % (self.conv.module_str, 'shortcut')
-
-    """@property
-    def get_active_parameter_size(self):
-        param_W = np.prod(self.conv.conv._W.shape)
-        param_b = np.prod(self.conv.conv._b.shape)
-        return param_W + param_b"""
-
     @staticmethod
     def build_from_config(config):
         conv_config = config['conv'] if 'conv' in config else config['mobile_inverted_conv']
         conv = set_layer_from_config(conv_config)
-        shortcut = set_layer_from_config(config['shortcut'])
+        shortcut = Mo.Identity()
         return ResidualBlock(conv, shortcut)
 
 
-class My2DLayer(Mo.Module):
+class ConvLayer(Mo.Sequential):
 
-    def __init__(self, in_channels, out_channels,
-                 use_bn=True, use_se=False, act_func='relu', dropout=0, ops_order='weight_bn_act'):
+    r"""Convolution-BatchNormalization(optional)-Activation layer.
 
-        super(My2DLayer, self).__init__()
+    Args:
+        in_channels (int): Number of convolution kernels (which is
+            equal to the number of input channels).
+        out_channels (int): Number of convolution kernels (which is
+            equal to the number of output channels). For example, to apply
+            convolution on an input with 16 types of filters, specify 16.
+        kernel (tuple of int, optional): Convolution kernel size. For
+            example, to apply convolution on an image with a 3 (height) by 5
+            (width) two-dimensional kernel, specify (3, 5). Defaults to (3, 3)
+        stride (tuple of int, optional): Stride sizes for
+            dimensions. Defaults to (1, 1).
+        dilation (tuple of int, optional): Dilation sizes for
+            dimensions. Defaults to (1, 1).
+        group (int, optional): Number of groups of channels.
+            Defaults to 1.
+        with_bias (bool, optional): If True, bias for Convolution is added.
+            Defaults to False.
+        use_bn (bool, optional): If True, BatchNormalization layer is added.
+            Defaults to True.
+        act_func (str, optional) Type of activation. Defaults to 'relu'.
+    """
 
-        self._scope_name = f'<my2dlayer at {hex(id(self))}>'
+    def __init__(self, in_channels, out_channels, kernel=(3, 3),
+                 stride=(1, 1), dilation=(1, 1), group=1, with_bias=False,
+                 use_bn=True, act_func='relu'):
         self._in_channels = in_channels
         self._out_channels = out_channels
-
-        self._use_bn = use_bn
-        self._use_se = use_se
-        self._act_func = act_func
-        self._dropout = dropout
-        self._ops_order = ops_order
-
-        modules = {}
-        # bn
-        if self._use_bn:
-            if self.bn_before_weight:
-                modules['bn'] = Mo.BatchNormalization(in_channels, 4)
-            else:
-                modules['bn'] = Mo.BatchNormalization(out_channels, 4)
-        else:
-            modules['bn'] = None
-        # activation
-        modules['act'] = build_activation(self._act_func)
-        # dropout
-        if self._dropout > 0:
-            modules['dropout'] = Mo.Dropout(self._dropout)
-        else:
-            modules['dropout'] = None
-        # weight
-        modules['weight'] = self.weight_op()
-
-        # add modules
-        module_dict = OrderedDict()
-        for op in self.ops_list:
-            if modules[op] is None:
-                continue
-            elif op == 'weight':
-                if modules['dropout'] is None:
-                    module_dict['dropout'] = modules['dropout']
-                for key in modules['weight']:
-                    module_dict[key] = modules['weight'][key]
-            else:
-                module_dict[op] = modules[op]
-        if self._use_se:
-            module_dict['se'] = SEModule(self._out_channels)
-        self._layer = Mo.Sequential(module_dict)
-
-    @property
-    def ops_list(self):
-        return self._ops_order.split('_')
-
-    @property
-    def bn_before_weight(self):
-        for op in self.ops_list:
-            if op == 'bn':
-                return True
-            elif op == 'weight':
-                return False
-        raise ValueError('Invalid ops_order: %s' % self._ops_order)
-
-    def weight_op(self):
-        raise NotImplementedError
-
-    def call(self, input):
-        for module in self._layer:
-            input = module(input)
-        return input
-
-    @property
-    def module_str(self):
-        raise NotImplementedError
-
-    @staticmethod
-    def build_from_config(config):
-        raise NotImplementedError
-
-
-class ConvLayer(My2DLayer):
-
-    def __init__(self, in_channels, out_channels,
-                 kernel=(3, 3), stride=(1, 1),
-                 dilation=(1, 1), group=1, with_bias=False,
-                 has_shuffle=False, use_se=False, use_bn=True, act_func='relu',
-                 dropout=0, ops_order='weight_bn_act'):
-
-        self._scope_name = f'<convlayer at {hex(id(self))}>'
-
         self._kernel = kernel
         self._stride = stride
         self._dilation = dilation
         self._group = group
         self._with_bias = with_bias
-        self._has_shuffle = has_shuffle
-        self._use_se = use_se
+        self._use_bn = use_bn
+        self._act_func = act_func
 
-        super(ConvLayer, self).__init__(in_channels, out_channels, use_bn, use_se, act_func, dropout, ops_order)
-
-    def weight_op(self):
         padding = get_same_padding(self._kernel)
         if isinstance(padding, int):
             padding *= self._dilation
@@ -189,58 +235,60 @@ class ConvLayer(My2DLayer):
             new_padding = (padding[0] * self._dilation[0], padding[1] * self._dilation[1])
             padding = tuple(new_padding)
 
-        weight_dict = OrderedDict({
-            'conv': Mo.Conv(
-                self._in_channels, self._out_channels, self._kernel, pad=padding,
-                stride=self._stride, dilation=self._dilation,
-                group=min_divisible_value(self._in_channels, self._group),
-                with_bias=self._with_bias)
-        })
-        if self._has_shuffle and self._group > 1:
-            print('shuffle not implemented yet')
+        module_dict = OrderedDict()
+        module_dict['conv'] = Mo.Conv(self._in_channels, self._out_channels, self._kernel,
+                                      pad=padding, stride=self._stride, dilation=self._dilation,
+                                      group=min_divisible_value(self._in_channels, self._group),
+                                      with_bias=self._with_bias)
+        if self._use_bn:
+            module_dict['bn'] = Mo.BatchNormalization(out_channels, 4)
+        module_dict['act'] = build_activation(act_func)
 
-        return weight_dict
+        super(ConvLayer, self).__init__(module_dict)
 
-    @property
-    def module_str(self):
-        if isinstance(self.kernel_size, int):
-            kernel_size = (self.kernel_size, self.kernel_size)
-        else:
-            kernel_size = self.kernel_size
-        if self.groups == 1:
-            if self.dilation > 1:
-                conv_str = '%dx%d_DilatedConv' % (kernel_size[0], kernel_size[1])
-            else:
-                conv_str = '%dx%d_Conv' % (kernel_size[0], kernel_size[1])
-        else:
-            if self.dilation > 1:
-                conv_str = '%dx%d_DilatedGroupConv' % (kernel_size[0], kernel_size[1])
-            else:
-                conv_str = '%dx%d_GroupConv' % (kernel_size[0], kernel_size[1])
-        conv_str += '_O%d' % self.out_channels
-        if self.use_se:
-            conv_str = 'SE_' + conv_str
-        conv_str += '_' + self.act_func.upper()
-        if self.use_bn:
-            if isinstance(self.bn, nn.GroupNorm):
-                conv_str += '_GN%d' % self.bn.num_groups
-            elif isinstance(self.bn, nn.BatchNorm2d):
-                conv_str += '_BN'
-        return conv_str
-
-    @staticmethod
     def build_from_config(config):
         return ConvLayer(**config)
 
+    def extra_repr(self):
+        return (f'in_channels={self._in_channels}, '
+                f'out_channels={self._out_channels}, '
+                f'kernel={self._kernel}, '
+                f'stride={self._stride}, '
+                f'dilation={self._dilation}, '
+                f'group={self._group}, '
+                f'with_bias={self._with_bias}, '
+                f'use_bn={self._use_bn}, '
+                f'act_func={self._act_func}, '
+                f'name={self._name}')
+
 
 class MBConvLayer(Mo.Module):
+
+    r"""The inverted layer with optional squeeze-and-excitation.
+
+    Args:
+        in_channels (int): Number of convolution kernels (which is
+            equal to the number of input channels).
+        out_channels (int): Number of convolution kernels (which is
+            equal to the number of output channels). For example, to apply
+            convolution on an input with 16 types of filters, specify 16.
+        kernel (tuple of int): Convolution kernel size. For
+            example, to apply convolution on an image with a 3 (height) by 5
+            (width) two-dimensional kernel, specify (3, 5). Defaults to (3, 3)
+        stride (tuple of int, optional): Stride sizes for dimensions.
+            Defaults to (1, 1).
+        expand_ratio (int): The expand ratio.
+        mid_channels (int): The number of features. Defaults to None.
+        act_func (str) Type of activation. Defaults to 'relu'.
+        use_se (bool, optional): If True, squeeze-and-expand module is used.
+            Defaults to False.
+        group (int, optional): Number of groups of channels.
+            Defaults to 1.
+    """
+
     def __init__(self, in_channels, out_channels,
                  kernel=(3, 3), stride=(1, 1), expand_ratio=6, mid_channels=None,
                  act_func='relu6', use_se=False, group=None):
-
-        super(MBConvLayer, self).__init__()
-        self._scope_name = f'<mbconvlayer at {hex(id(self))}>'
-
         self._in_channels = in_channels
         self._out_channels = out_channels
         self._kernel = kernel
@@ -283,204 +331,63 @@ class MBConvLayer(Mo.Module):
             ('bn', Mo.BatchNormalization(out_channels, 4))
         ]))
 
-    def call(self, input):
+    def call(self, x):
         if self.inverted_bottleneck:
-            input = self.inverted_bottleneck(input)
-        x = self.depth_conv(input)
+            x = self.inverted_bottleneck(x)
+        x = self.depth_conv(x)
         x = self.point_linear(x)
         return x
-
-    @property
-    def module_str(self):
-        if self._mid_channels is None:
-            expand_ratio = self._expand_ratio
-        else:
-            expand_ratio = self._mid_channels // self._in_channels
-        layer_str = '%dx%d_MBConv%d_%s' % (self._kernel, self._kernel, expand_ratio, self.act_func.upper())
-        if self._use_se:
-            layer_str = 'SE_' + layer_str
-        layer_str += '_O%d' % self._out_channels
-        if self._group is not None:
-            layer_str += '_G%d' % self._group
-        if isinstance(self.point_linear.bn, nn.GroupNorm):
-            layer_str += '_GN%d' % self.point_linear.bn.num_groups
-        elif isinstance(self.point_linear.bn, nn.BatchNorm2d):
-            layer_str += '_BN'
-
-        return layer_str
 
     @staticmethod
     def build_from_config(config):
         return MBConvLayer(**config)
 
-
-class ResNetBottleneckBlock(Mo.Module):
-    def __init__(self, in_channels, out_channels,
-                 kernel=(3, 3), stride=(1, 1), expand_ratio=0.25, mid_channels=None, act_func='relu', group=1,
-                 downsample_mode='avgpool_conv'):
-
-        super(ResNetBottleneckBlock, self).__init__()
-        self._scope_name = f'<resnetbottleneclblock at {hex(id(self))}>'
-        self._in_channels = in_channels
-        self._out_channels = out_channels
-
-        self._kernel = kernel
-        self._stride = stride
-        self._expand_ratio = expand_ratio
-        self._mid_channels = mid_channels
-        self._act_func = act_func
-        self._group = group
-
-        self._downsample_mode = downsample_mode
-
-        if self._mid_channels is None:
-            feature_dim = round(self._out_channels * self._expand_ratio)
-        else:
-            feature_dim = self._mid_channels
-
-        feature_dim = make_divisible(feature_dim, MyNetwork.CHANNEL_DIVISIBLE)
-        self._mid_channels = feature_dim
-
-        self.conv1 = Mo.Sequential(OrderedDict([
-            ('conv', Mo.Conv(self._in_channels, feature_dim, (1, 1), pad=(0, 0), stride=(1, 1), with_bias=False)),
-            ('bn', Mo.BatchNormalization(feature_dim, 4)),
-            ('act', build_activation(self._act_func)),
-        ]))
-
-        pad = get_same_padding(self._kernel)
-        self.conv2 = Mo.Sequential(OrderedDict([
-            ('conv', Mo.Conv(feature_dim, feature_dim, kernel, pad=pad, stride=stride, group=group, with_bias=False)),
-            ('bn', Mo.BatchNormalization(feature_dim, 4)),
-            ('act', build_activation(self._act_func)),
-        ]))
-
-        self.conv3 = Mo.Sequential(OrderedDict([
-            ('conv', Mo.Conv(feature_dim, self._out_channels, (1, 1), pad=(0, 0), stride=(1, 1), with_bias=False)),
-            ('bn', Mo.BatchNormalization(self._out_channels, 4)),
-        ]))
-
-        if stride == (1, 1) and in_channels == out_channels:
-            self.downsample = IdentityLayer(in_channels, out_channels)
-        elif self._downsample_mode == 'conv':
-            self.downsample = Mo.Sequential(OrderedDict([
-                ('conv', Mo.Conv(in_channels, out_channels, (1, 1), pad=(0, 0), stride=stride, with_bias=False)),
-                ('bn', Mo.BatchNormalization(out_channels, 4)),
-            ]))
-        elif self._downsample_mode == 'avgpool_conv':
-            self.downsample = Mo.Sequential(OrderedDict([
-                ('avg_pool', Mo.AvgPool(kernel=stride, stride=stride, pad=(0, 0))),
-                ('conv', Mo.Conv(in_channels, out_channels, (1, 1), pad=(0, 0), stride=(1, 1), with_bias=False)),
-                ('bn', Mo.BatchNormalization(out_channels, 4)),
-            ]))
-        else:
-            raise NotImplementedError
-
-        self.final_act = build_activation(self._act_func)
-
-    def call(self, x):
-
-        residual = self.downsample(x)
-
-        x = self.conv1(x)
-        x = self.conv2(x)
-        x = self.conv3(x)
-
-        x = x + residual
-        x = self.final_act(x)
-
-        return x
+    def extra_repr(self):
+        return (f'in_channels={self._in_channels}, '
+                f'out_channels={self._out_channels}, '
+                f'kernel={self._kernel}, '
+                f'stride={self._stride}, '
+                f'expand_ratio={self._expand_ratio}, '
+                f'mid_channels={self._mid_channels}, '
+                f'act_func={self._act_func}, '
+                f'use_se={self._use_se}, '
+                f'group={self._group} ')
 
 
-class IdentityLayer(My2DLayer):
-    def __init__(self, in_channels, out_channels,
-                 use_bn=False, use_se=False, act_func=None, dropout=0, ops_order='weight_bn_act'):
-        super(IdentityLayer, self).__init__(in_channels, out_channels, use_bn, use_se, act_func, dropout, ops_order)
+class LinearLayer(Mo.Sequential):
 
-        self._scope_name = f'<identitylayer at {hex(id(self))}>'
+    r"""Affine, or fully connected layer with dropout.
 
-    def weight_op(self):
-        return None
+    Args:
+        in_features (int): The size of each input sample.
+        in_features (int): The size of each output sample.
+        with_bias (bool): Specify whether to include the bias term.
+            Defaults to True.
+        drop_rate (float, optional): Dropout ratio applied to parameters.
+            Defaults to 0.
+    """
 
-    @property
-    def module_str(self):
-        return 'Identity'
-
-    @staticmethod
-    def build_from_config(config):
-        return IdentityLayer(**config)
-
-
-class LinearLayer(Mo.Module):
-    def __init__(self, in_features, out_features, bias=True,
-                 use_bn=False, act_func=None, dropout=0, ops_order='weight_bn_act'):
-
-        super(LinearLayer, self).__init__()
-
-        self._scope_name = f'<linearlayer at {hex(id(self))}>'
+    def __init__(self, in_features, out_features, bias=True, drop_rate=0):
         self._in_features = in_features
         self._out_features = out_features
         self._bias = bias
+        self._drop_rate = drop_rate
 
-        self._use_bn = use_bn
-        self._act_func = act_func
-        self._dropout = dropout
-        self._ops_order = ops_order
-
-        modules = {}
-        # batch norm
-        if self._use_bn:
-            if self.bn_before_weight:
-                modules['bn'] = Mo.BatchNormalization(in_features, 2)
-            else:
-                modules['bn'] = Mo.BatchNormalization(out_features, 2)
-        else:
-            modules['bn'] = None
-        # activation
-        modules['act'] = build_activation(self._act_func)
-        # dropout
-        if self._dropout > 0:
-            modules['dropout'] = Mo.Dropout(self._dropout)
-        else:
-            modules['dropout'] = None
-        # linear
-        modules['weight'] = {
-            'linear': Mo.Linear(self._in_features, self._out_features, bias=self._bias)}
-
-        # add modules
-        module_dict = OrderedDict()
-        for op in self.ops_list:
-            if modules[op] is None:
-                continue
-            elif op == 'weight':
-                if modules['dropout'] is not None:
-                    module_dict['dropout'] = modules['dropout']
-                for key in modules['weight']:
-                    module_dict[key] = modules['weight'][key]
-            else:
-                module_dict[op] = modules[op]
-        self.linearlayer = Mo.Sequential(module_dict)
-
-    @property
-    def ops_list(self):
-        return self._ops_order.split('_')
-
-    @property
-    def bn_before_weight(self):
-        for op in self.ops_list:
-            if op == 'bn':
-                return True
-            elif op == 'weight':
-                return False
-        raise ValueError('Invalid ops_order: %s' % self.ops_order)
-
-    def call(self, input):
-        for module in self.linearlayer:
-            input = module(input)
-        return input
+        super(LinearLayer, self).__init__(OrderedDict({
+            'dropout': Mo.Dropout(self._drop_rate),
+            'linear': Mo.Linear(self._in_features, self._out_features, bias=self._bias),
+        }))
 
     @staticmethod
     def build_from_config(config):
         return LinearLayer(**config)
+
+    def extra_repr(self):
+        return (f'in_channels={self._in_channels}, '
+                f'out_channels={self._out_channels}, '
+                f'bias={self._bias}, '
+                f'drop_rate={self._drop_rate}, '
+                f'name={self._name} ')
 
 
 def build_activation(act_func, inplace=False):
@@ -489,22 +396,15 @@ def build_activation(act_func, inplace=False):
     elif act_func == 'relu6':
         return Mo.ReLU6()
     elif act_func == 'h_swish':
-        return Hswish()
+        return Mo.Hswish()
     elif act_func is None or act_func == 'none':
         return None
     else:
         raise ValueError('do not support: %s' % act_func)
 
 
-def adjust_bn_according_to_idx_ver1(bn, idx):
-    bn._beta = F.gather(bn._beta, idx, 1)
-    bn._gamma = F.gather(bn._gamma, idx, 1)
-    bn._mean = F.gather(bn._mean, idx, 1)
-    bn._var = F.gather(bn._var, idx, 1)
-
-
-def adjust_bn_according_to_idx(bn, idx):
+"""def adjust_bn_according_to_idx(bn, idx):
     bn._beta.d = np.stack([bn._beta.d[:, i, :, :] for i in idx], axis=1)
     bn._gamma.d = np.stack([bn._gamma.d[:, i, :, :] for i in idx], axis=1)
     bn._mean.d = np.stack([bn._mean.d[:, i, :, :] for i in idx], axis=1)
-    bn._var.d = np.stack([bn._var.d[:, i, :, :] for i in idx], axis=1)
+    bn._var.d = np.stack([bn._var.d[:, i, :, :] for i in idx], axis=1)"""
