--- nnabla_nas/contrib/classification/ofa/utils.py
@@ -1,70 +1,0 @@
-# Functions for OFA (original implementation from pytorch code)
-
-import numpy as np
-
-import nnabla.functions as F
-
-
-def label_smoothing_loss(pred, label, label_smoothing=0.1):
-    loss = F.softmax_cross_entropy(pred, label)
-    if label_smoothing <= 0:
-        return loss
-    return (1 - label_smoothing) * loss - label_smoothing \
-        * F.mean(F.log_softmax(pred), axis=1, keepdims=True)
-
-
-def min_divisible_value(n1, v1):
-    """ make sure v1 is divisible by n1, otherwise decrease v1 """
-    if v1 >= n1:
-        return n1
-    while n1 % v1 != 0:
-        v1 -= 1
-    return v1
-
-
-def val2list(val, repeat_time=1):
-    if isinstance(val, list) or isinstance(val, np.ndarray):
-        return val
-    elif isinstance(val, tuple):
-        return list(val)
-    else:
-        return [val for _ in range(repeat_time)]
-
-
-def make_divisible(v, divisor=8, min_val=None):
-    """
-    This function is taken from the original tf repo.
-    It ensures that all layers have a channel number that is divisible by 8
-    It can be seen here:
-    https://github.com/tensorflow/models/blob/master/research/slim/nets/mobilenet/mobilenet.py
-    :param v:
-    :param divisor:
-    :param min_val:
-    :return:
-    """
-    if min_val is None:
-        min_val = divisor
-    new_v = max(min_val, int(v + divisor / 2) // divisor * divisor)
-    # Make sure that round down does not go down by more than 10%.
-    if new_v < 0.9 * v:
-        new_v += divisor
-    return new_v
-
-
-def get_same_padding(kernel_size):
-    if isinstance(kernel_size, tuple):
-        assert len(kernel_size) == 2, 'invalid kernel size: %s' % kernel_size
-        p1 = get_same_padding(kernel_size[0])
-        p2 = get_same_padding(kernel_size[1])
-        return p1, p2
-    assert isinstance(kernel_size, int), 'kernel size should be either `int` or `tuple`'
-    assert kernel_size % 2 > 0, 'kernel size should be odd number'
-    return kernel_size // 2
-
-
-def sub_filter_start_end(kernel_size, sub_kernel_size):
-    center = kernel_size // 2
-    dev = sub_kernel_size // 2
-    start, end = center - dev, center + dev + 1
-    assert end - start == sub_kernel_size
-    return start, end
